# statistics: add branch check for BinarySearchRemoveVal